feat(api): project status
