+++ tests/api/test_project.py
@@ -22,6 +22,8 @@ import os
 import pytest
 
 from renku.ui.api import Project
+from renku.ui.cli import cli
+from tests.utils import format_result_exception, write_and_commit_file
 
 
 @pytest.mark.parametrize("sub_path", [".", "src", "src/notebooks"])
@@ -47,7 +49,7 @@ def test_get_project_multiple(client):
 
 
 def test_get_or_create_project(client):
-    """Test getting Project context or creating one yileds similar results."""
+    """Test getting Project context or creating one yields similar results."""
     with Project() as project_1:
         pass
 
@@ -62,3 +64,28 @@ def test_get_project_outside_a_renku_project(directory_tree):
 
     with Project() as project:
         assert project.client is not None
+
+
+def test_status(runner, client):
+    """Test status check."""
+    source = client.path / "source.txt"
+    output = client.path / "data" / "output.txt"
+
+    write_and_commit_file(client.repository, source, "content")
+
+    result = runner.invoke(cli, ["run", "cp", source, output])
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    result = runner.invoke(cli, ["run", "cat", "--no-output", source])
+    assert 0 == result.exit_code, format_result_exception(result)
+
+    write_and_commit_file(client.repository, source, "new content")
+
+    result = Project().status()
+    stale_generations, stale_activities, modified_inputs, deleted_inputs = result
+
+    assert (os.path.relpath(output), {os.path.relpath(source)}) in stale_generations.items()
+    assert "source.txt" in modified_inputs
+    assert 1 == len(stale_activities)
+    assert "/activities/" in list(stale_activities.keys())[0]
+    assert 0 == len(deleted_inputs)
